refactor ui inspector trace sections
